Finallized everything and nothing is broken
