+++ Canvas2.0/Canvas2.0.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33213.308
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Canvas2.0", "Canvas2.0.csproj", "{352E4C4E-F200-4DC4-8F05-6BEA5D61E948}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Objects", "..\Objects\Objects.csproj", "{A081E17F-E999-4617-9BF5-1FA17643E507}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{352E4C4E-F200-4DC4-8F05-6BEA5D61E948}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{352E4C4E-F200-4DC4-8F05-6BEA5D61E948}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{352E4C4E-F200-4DC4-8F05-6BEA5D61E948}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{352E4C4E-F200-4DC4-8F05-6BEA5D61E948}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A081E17F-E999-4617-9BF5-1FA17643E507}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A081E17F-E999-4617-9BF5-1FA17643E507}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A081E17F-E999-4617-9BF5-1FA17643E507}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A081E17F-E999-4617-9BF5-1FA17643E507}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {09C62F53-0CFD-4C28-AAAD-EDBAE1BBC16D}
+	EndGlobalSection
+EndGlobal
