+++ Library.TaskManagement/Models/ToDo.cs
@@ -15,5 +15,11 @@
             return Display;
         }
 
+        public ToDo(Item i )
+        {
+            Name = i.Name;
+            Description= i.Description;
+        }
+
     }
 }
