+++ Objects/Models/ContentItem.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Objects.Models
+{
+    public class ContentItem : Item
+    {
+        public string contentPath { get; set; }
+    }
+}
