+++ Objects/Models/Item.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Objects.Models
+{
+    public class Item
+    {
+        public string Name { get; set; }
+        public string Description { get; set; }
+    }
+}
